requirements.txt
streamlit==1.36.0
langchain==0.1.16
langchain-community==0.0.32
langchain-core==0.1.45
langchain-groq==0.1.3
pypdf==4.2.0
faiss-cpu==1.8.0
sentence-transformers==2.7.0
python-dotenv==1.0.1